+++ sotodlib/stimulator/stimulator.py
@@ -9,6 +9,7 @@ import ruptures as rpt
 from sotodlib import tod_ops, core
 from sotodlib.io import hkdb
 from sotodlib.stimulator.plot_stimulator import plot_hkdata, plot_tod
+from sotodlib.stimulator.utils_stimulator import func_sines, func_response_amplitude, func_response_phase, func_response_phase_with_dt  
 
 CHOPPING_FREQS = {'f1':6, 'f2':15,'f3':33,'f4':63,'f5':93,'f6':123,'f7':147}
 
@@ -869,84 +870,6 @@ def get_fit_params(cal_type):
         params_base['fit_phase__free']   .add('dt',value=0.125*1e-3,min=-3e-3,max=3e-3)
 
     return model,params_base
-
-
-def func_sines(t, a0, a1, a2, a3, a4, a5, a6, t0, t1, t2, t3, t4, t5, t6):
-    """
-    Define fitting function.
-    
-    Args:
-        t: time or timing_frac of stimulator signal.
-        a0-6: Amplitude of sin function for first to seventh frequency signal
-        t0-6: Timing offset of sin function for first to seventh frequency signal
-    
-    Return:
-        y: Value of fitting function.
-    """
-    y = (a0*np.sin(1*(t-t0)*2*np.pi)
-         + a1*np.sin(2*(t-t1)*2*np.pi)
-         + a2*np.sin(3*(t-t2)*2*np.pi)
-         + a3*np.sin(4*(t-t3)*2*np.pi)
-         + a4*np.sin(5*(t-t4)*2*np.pi)
-         + a5*np.sin(6*(t-t5)*2*np.pi)
-         + a6*np.sin(7*(t-t6)*2*np.pi))
-    
-    return y
-
-
-def func_response_amplitude(f, tau, a):
-    """
-    Detector response function of amplitude.
-    
-    Args:
-        f: Chopping frequency
-        tau: time constant of a detector in [s]
-        a: Amplitude of sin function for '0' frequency signal
-
-    Return:
-        y: Value of fitting function.
-    """
-    y = a /np.sqrt(1+(2*np.pi*f*tau)**2)
-    return y
-
-
-def func_response_phase(f, tau, theta_geo):
-    """
-    Detector response function of phase.
-    
-    Args:
-        f: Chopping frequency [Hz]
-        tau: time constant of a detector in [s]
-        theta_geo: Offset of phase delay [deg]
-        theta: Phase delay of stimulator signal [deg]
-
-    Return:
-        theta: Value of fitting function.
-    """
-    theta = np.arctan(-2*np.pi*f*tau)*(180/np.pi) + theta_geo
-    return theta
-
-
-def func_response_phase_with_dt(f, tau, theta_geo, dt):
-    """
-    Detector response function of phase.
-    
-    Args:
-        f: Chopping frequency [Hz]
-        tau: time constant of a detector in [s]
-        theta: Phase delay of stimulator signal [deg]
-        theta_geo: Offset of phase delay due to hardware effect, geo = geometry [deg]
-        theta_dt: Offset of phase delay due to readout issue [deg], theta_dt*(pi/180) =  -delta_t*2pi*f
-        dt: Time difference due to wrong time stamps
-
-    Return:
-        theta: Value of fitting function.
-    """
-    theta_dt = -dt*2*np.pi*f *(180/np.pi)
-    theta = np.arctan(-2*np.pi*f*tau)*(180/np.pi) + theta_geo + theta_dt
-    return theta
-
-
 
 
 def fill_data(aman, coadd_data, fit_result, n_bins, cal_type):
